CARSOLD #210 | 1) Simplified Offer entity to work with followedOffers properly instead of follows field 2) Updated FunctionalityServiceImpl methods 3) Updated delete Offer/User methods 4) Adjusted tests 5) Semantic changes to Navbar and MidBar

+++ CARSOLD/src/main/java/org/gontar/carsold/Domain/Model/Offer/OfferStatsDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OfferStatsDto {
-    private Integer views;
-    private Integer follows;
+    private Long views;
+    private Long follows;
 }
